+++ HISTORY.md
@@ -1,8 +1,10 @@
 # Version 0.0.8
+- Added support for 'arch.univariate.base.ARCHModelResult'
 - Simplified and extended `calculate_residuals_statistics()`
 - Added covariance type and fixed effects table to `linearmodels.panel.result.PanelEffectsresult` print result
 - Moved default values to function definitions
 - Updated required install to pandas only
+- Added support for `arch.univariate.base.ARCHModelResult`
 
 # Version 0.0.7
 - Moved type checks to functions and globals